Test Bug mit den Datentypen der callback signale auf der Arm HW zu fixen

+++ src/gui/gtk3/view_calibrator.rs
@@ -120,19 +120,21 @@ pub fn launch(sensor_type: SensorType, builder: &gtk::Builder, kombisensor: &Arc
 
     // Zurueck Action
     button_calibrator_cancel.connect_clicked(move |_| {
+        use gui::gtk3::libc::c_ulong;
+
         unsafe {
             if gobject_ffi::g_signal_handler_is_connected(button_messpunkt_nullgas.to_glib_none().0, id_button_messpunkt_nullgas) == 1 {
-                gobject_ffi::g_signal_handler_disconnect(button_messpunkt_nullgas.to_glib_none().0, id_button_messpunkt_nullgas);
+                gobject_ffi::g_signal_handler_disconnect(button_messpunkt_nullgas.to_glib_none().0, id_button_messpunkt_nullgas as c_ulong);
             }
         }
         unsafe {
             if gobject_ffi::g_signal_handler_is_connected(button_messpunkt_messgas.to_glib_none().0, id_button_messpunkt_messgas) == 1 {
-                gobject_ffi::g_signal_handler_disconnect(button_messpunkt_messgas.to_glib_none().0, id_button_messpunkt_messgas);
+                gobject_ffi::g_signal_handler_disconnect(button_messpunkt_messgas.to_glib_none().0, id_button_messpunkt_messgas as c_ulong);
             }
         }
         unsafe {
             if gobject_ffi::g_signal_handler_is_connected(button_calibrator_save.to_glib_none().0, id_button_calibrator_save) == 1 {
-                gobject_ffi::g_signal_handler_disconnect(button_calibrator_save.to_glib_none().0, id_button_calibrator_save);
+                gobject_ffi::g_signal_handler_disconnect(button_calibrator_save.to_glib_none().0, id_button_calibrator_save as c_ulong);
             }
         }
         stack_main.set_visible_child(&box_index_view);
